# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "U-ED News API Documentation",
        "description": "A simple API made with Laravel 8",
        "contact": {
            "email": "work.jerson.baltazar@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "description": "Returns list of article data",
                "operationId": "indexArticle",
                "parameters": [
                    {
                        "name": "header",
                        "in": "query",
                        "description": "Filter by header of the article",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_name",
                        "in": "query",
                        "description": "Filter by the name of article's category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "creator_full_name",
                        "in": "query",
                        "description": "Filter by the full name of article's creator",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags_name",
                        "in": "query",
                        "description": "Filter by the name of article's tags",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status of article 1 = Active, 0 = Inactive",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "15"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Store new article",
                "description": "Returns article data",
                "operationId": "storeArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreArticleRequestModel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/articles/{article}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Show certain article",
                "description": "Returns an article data",
                "operationId": "showArticle",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update existing article",
                "description": "Returns updated article data",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateArticleRequestModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete existing Article",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteArticle",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "Returns token and user details",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestModel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequestModel": {
                "title": "Login request",
                "description": "Login request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email of the user",
                        "type": "string",
                        "example": "super.admin@yopmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password of the user",
                        "type": "string",
                        "example": "Password@123"
                    }
                },
                "type": "object"
            },
...
            "StoreArticleRequestModel": {
                "title": "Store Article request",
                "description": "Store Article request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "header": {
                        "title": "header",
                        "description": "header of the new article",
                        "type": "string",
                        "example": "A nice article"
                    },
                    "body": {
                        "title": "body",
                        "description": "body of the new article",
                        "type": "string",
                        "example": "<p>Content here</p>"
                    },
                    "video_path": {
                        "title": "video_path",
                        "description": "body of the new article",
                        "type": "string",
                        "example": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                    },
                    "thumbnail_path": {
                        "title": "thumbnail_path",
                        "description": "body of the new article",
                        "type": "string",
                        "example": "https://i.ytimg.com/an_webp/dQw4w9WgXcQ/mqdefault_6s.webp?du=3000&sqp=CLisy4cG&rs=AOn4CLDj6lo3qNP4APhTvCFcJLlAEKlYkw"
                    },
                    "article_category_id": {
                        "title": "article_category_id",
                        "description": "body of the new article",
                        "type": "integer",
                        "example": "1"
                    },
                    "tag_ids": {
                        "title": "tag_ids",
                        "description": "body of the new article",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateArticleRequestModel": {
                "title": "Update Article request",
                "description": "Update Article request body data",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/StoreArticleRequestModel"
                    },
                    {
                        "required": [
                            "name"
                        ],
                        "properties": {
                            "status": {
                                "title": "status",
                                "description": "status of the new article 1 = Active, 0 = Inactive",
                                "type": "string",
                                "example": "1"
                            },
                            "view_count": {
                                "title": "view_count",
                                "description": "view_count of the new article",
                                "type": "string",
                                "example": "100"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}